Adequate the frontend to make requests to create, get, edit and delete habits and habitLogs adding a heatmap interface where users can select the day which they have completed the habit, therefore sending a habit log with a check(1 or 0) or a rational value

+++ habittracker-backend/habittracker/src/main/java/com/example/habittracker/dto/HabitDetailsDTO.java
@@ -4,13 +4,13 @@ import java.time.LocalDate;
 import java.util.Map;
 
 public record HabitDetailsDTO(
-        Long id,
-        String name,
-        String unit,
-        String colorCode,
-        Integer goal,
+                Long id,
+                String name,
+                String unit,
+                String colorCode,
+                Double goal,
 
-        // Chave = Data (2025-11-22)
-        // Valor = Quanto fez (1 ou 1000ml)
-        Map<LocalDate, Integer> history) {
+                // Chave = Data (2025-11-22)
+                // Valor = Quanto fez (1 ou 1000ml)
+                Map<LocalDate, Double> history) {
 }
