+++ habittracker-backend/habittracker/src/main/java/com/example/habittracker/dto/HabitLogRequestDTO.java
@@ -4,9 +4,9 @@ import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 
 public record HabitLogRequestDTO(
-        @NotNull(message = "Date is required") LocalDate date,
+                @NotNull(message = "Date is required") LocalDate date,
 
-        // Se for null, o service assume 1 (feito).
-        // Se tiver valor (ex: 500ml), usa o valor.
-        Integer value) {
+                // Se for null, o service assume 1 (feito).
+                // Se tiver valor (ex: 500ml), usa o valor.
+                Double value) {
 }
