+++ habittracker-backend/habittracker/src/main/java/com/example/habittracker/dto/HabitLogResponseDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 public record HabitLogResponseDTO(
         Long id,
         LocalDate date,
-        Integer value) {
+        Double value) {
     public HabitLogResponseDTO(HabitLog log) {
         this(log.getId(), log.getDate(), log.getValue());
     }
