+++ habittracker-backend/habittracker/src/main/java/com/example/habittracker/dto/HabitRequestDTO.java
@@ -4,9 +4,9 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
 public record HabitRequestDTO(
-        @NotBlank(message = "Name is mandatory") @Size(min = 1, max = 100) String name,
+                @NotBlank(message = "Name is mandatory") @Size(min = 1, max = 100) String name,
 
-        String unit,
-        String colorCode,
-        Integer goal) {
+                String unit,
+                String colorCode,
+                Double goal) {
 }
